refactor: refactor SuccessPage component
