+++ src/components/Main/Main.css
@@ -1,7 +1,7 @@
-.container {
+/* .container {
   padding: 0 10ch;
   justify-content: center;
-}
+} */
 
 .burger {
   display: flex;
@@ -12,9 +12,7 @@
 }
 
 @media screen and (max-width: 1050px) {
-
   .container {
     padding: 0 20px;
   }
-
 }
