--- src/components/PagesSuccess/PagesSuccess.css
@@ -1,13 +1,0 @@
-.succespage-container {
-    width: 100%;
-    height: 90vh;
-    display: flex;
-    justify-content: center;
-}
-
-.succespage-registration {
-    display: block;
-    text-align: center;
-    margin-top: 100px;
-    max-width: 620px;
-}
