+++ src/components/RoutingRoad/RoutingRoad.css
@@ -1,0 +1,3 @@
+.routing-road {
+  padding: 0 10ch;
+}
