+++ src/components/SuccessPage/SuccessPage.css
@@ -1,0 +1,13 @@
+.success-page__container {
+  width: 100%;
+  height: 90vh;
+  display: flex;
+  justify-content: center;
+}
+
+.success-page__registration {
+  display: block;
+  text-align: center;
+  margin-top: 100px;
+  max-width: 620px;
+}
