+++ src/components/TableForSale/TableForSale.css
@@ -27,3 +27,8 @@
 .basket-right__total {
   margin-top: 10%;
 }
+
+/* новое */
+.table-selling {
+  padding: 0 10ch;
+}
